requirements.txt
opencv-contrib-python>=4.5.0
numpy>=1.26.0
pandas>=1.5.0
Pillow>=9.3.0
